requirements.txt
pymongo
requests
python-dotenv
tenacity
click
schedule
tqdm
logging-utilities
